Errors in the parsing steps fix later.
The parsing step is currently not functioning. Right now it likes to break saying that my dataframe isn't there so i'm not quite sure how to fix that. Possible solutions are to start printing out the parsing as it goes and to change the storage system. If all else fails can sacrafice RAM and go back to original system.