Thermals Completed
I've completed the thermal profile generation which concludes the Skew-T analysis. Last part of the project is to make a scatter plot over a map proj. I advise stealing the code you had with sam and fixing it up a little more. Once that is done save that as a file to the git repo which can be called at any time in the future. This should have lat and lon values so that any .xlsx files put in with lat lon values.