Skew-T and Thermals
Last part is the scatter plot over the NA map. Take the code you did with sam and push it to a .pkl file so you can call it at any time. Once that is done get noah's .xlsx file and parse like you did from cole's. Other than that should be done.